refactor(core): extract registerKeyboardListeners
Pull the keydown + keyup listener bodies (210 lines combined) out of
init() into a single registerKeyboardListeners() helper. They share
tightly-coupled state machines (hold-timer, copy-feedback cooldown,
activation-key parser, toolbar-menu dismiss, prompt-mode, arrow-nav,
space-drag repositioning) so they live in one focused module.

Removed 14 now-unused imports/destructures from index.tsx.

index.tsx: 1166 -> 974 (-192). Sub-1000! 48 keyboard+shift specs green.
